requirements.txt
langchain==0.1.0
langchain-core==0.1.0
langchain-google-genai==1.0.0

google-generativeai==0.3.0

pydantic==2.5.0
python-dotenv==1.0.0

# 基础工具
pathlib
json
glob
datetime
typing
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.0.0
nltk==3.8.1
spacy==3.7.2
python-docx==0.8.11 